Fix the Clang-Tidy warnings flagged by Android Studio (#30)
All 66 warnings from the default inspection profile, no behavior change:
- narrowing int/uint32_t-to-float conversions: explicit (float) casts
- uninitialized record types: default member initializers on the small
  POD structs (Alien, Bullet, Bomb, Saucer, PowerUp, Particle, Explosion,
  Star, Pointer, DrawCmd) and {} on the fread buffers / timespec
- fireHeld: loop replaced with std::any_of
- mergeHighScore: index loop replaced with a range-based for
- emit and the draw helpers made const, numDigits made static

Intentional identifiers flagged as typos (frand, SINV, octo, ...) are
left as-is on purpose.

+++ app/src/main/cpp/test/game_test.cpp
@@ -164,7 +164,7 @@ TEST(March, StaysWithinSideMargins) {
             // Centers reverse at asp - margin - halfWidth = asp - 0.053;
             // 0.05 leaves slack for a single frame of overshoot.
             ASSERT_LT(fabsf(g.alienXForTest(a)), aspect() - 0.05f)
-                << "alien " << a << " out of margin after " << (i + 1) * 0.1f << " s";
+                << "alien " << a << " out of margin after " << (float)(i + 1) * 0.1f << " s";
         }
     }
 }
